chore(plugins): rename shims to `store.ts`
