+++ packages/plugins/drag-n-drop/src/index.ts
@@ -1,3 +1,3 @@
-import './store-shims.d'
+import './store'
 export * from './types'
 export * from './dnd'
